// src/api/storage.ts
export enum Storage {
  prefix = 'sim',
  token = 'token',
  refresh_token = 'refresh_token',
  theme = 'theme',
  history = 'history',
  stocks = 'stocks',
  keywords = 'keywords',
  search_history_keywords = 'search_history_keywords',
  search_history_object = 'search_history_object',
  info_membership = 'info_membership',
  portfolio_opt_current = 'portfolio_opt_current',
  user_info = 'user_info',
  current_screener = 'current_screener',
  referral_code = 'referral_code',
  current_user = 'current_user',
  course_review = 'course_review',
  watchlist_select = 'watchlist_select',
  vn_index_chart = 'vn_index_chart',
  vn_index_detail = 'vn_index_detail',
  vn_index_data = 'vn_index_data',
  campaint_normal_popup = 'campaint_normal_popup',
  campaint_slide_in_popup = 'campaint_slide_in_popup',
  current_campaint = 'current_campaint',
  sale_off_slide_in = 'sale_off_slide_in',
  is_open_leo = 'is_open_leo',
  current_nav_chart = 'current_nav_chart',
  current_summary_tab_chart = 'current_summary_tab_chart',
  is_show_full_summary_chart = 'is_show_full_summary_chart',
  lock_device = 'lock_device',
  current_chart_ticker = 'current_chart_ticker',
  analysis_tab = 'analysis_tab',
  ga_tracking = 'ga_tracking',
  pre_order_popup = 'pre_order_popup',
  popup = 'popup',
  notice = 'notice',
  mode = 'mode',
}